+++ src/main/java/net/imagej/ops/convert/converters/IntArrayToDimensionsConverter.java
@@ -1,4 +1,4 @@
-package net.imagej.ops.create.helper;
+package net.imagej.ops.convert.converters;
 
 import net.imglib2.Dimensions;
 import net.imglib2.FinalInterval;
@@ -9,15 +9,21 @@ import org.scijava.convert.Converter;
 import org.scijava.plugin.Plugin;
 
 @Plugin(type = Converter.class)
-public class LongArrayToDimensionsConverter extends
-		AbstractConverter<long[], Dimensions> implements
-		Converter<long[], Dimensions> {
+public class IntArrayToDimensionsConverter extends
+		AbstractConverter<int[], Dimensions> implements
+		Converter<int[], Dimensions> {
 
 	@SuppressWarnings("unchecked")
 	@Override
 	public <T> T convert(Object src, Class<T> dest) {
-		long[] input = (long[]) src;
-		return (T) new FinalInterval(input);
+
+		int[] input = (int[]) src;
+		long[] convert = new long[input.length];
+		for (int i = 0; i < input.length; i++) {
+			convert[i] = input[i];
+		}
+
+		return (T) new FinalInterval(convert);
 	}
 
 	@Override
@@ -26,14 +32,13 @@ public class LongArrayToDimensionsConverter extends
 	}
 
 	@Override
-	public Class<long[]> getInputType() {
-		return long[].class;
+	public Class<int[]> getInputType() {
+		return int[].class;
 	}
 
 	@Override
 	public boolean supports(ConversionRequest request) {
-
-		return request.sourceClass() == long[].class
+		return request.sourceClass() == int[].class
 				&& request.destType() == Dimensions.class;
 	}
 }
