+++ src/main/java/net/imagej/ops/convert/converters/LongArrayToDimensionsConverter.java
@@ -1,4 +1,4 @@
-package net.imagej.ops.create.helper;
+package net.imagej.ops.convert.converters;
 
 import net.imglib2.Dimensions;
 import net.imglib2.FinalInterval;
@@ -9,21 +9,15 @@ import org.scijava.convert.Converter;
 import org.scijava.plugin.Plugin;
 
 @Plugin(type = Converter.class)
-public class IntArrayToDimensionsConverter extends
-		AbstractConverter<int[], Dimensions> implements
-		Converter<int[], Dimensions> {
+public class LongArrayToDimensionsConverter extends
+		AbstractConverter<long[], Dimensions> implements
+		Converter<long[], Dimensions> {
 
 	@SuppressWarnings("unchecked")
 	@Override
 	public <T> T convert(Object src, Class<T> dest) {
-
-		int[] input = (int[]) src;
-		long[] convert = new long[input.length];
-		for (int i = 0; i < input.length; i++) {
-			convert[i] = input[i];
-		}
-
-		return (T) new FinalInterval(convert);
+		long[] input = (long[]) src;
+		return (T) new FinalInterval(input);
 	}
 
 	@Override
@@ -32,13 +26,14 @@ public class IntArrayToDimensionsConverter extends
 	}
 
 	@Override
-	public Class<int[]> getInputType() {
-		return int[].class;
+	public Class<long[]> getInputType() {
+		return long[].class;
 	}
 
 	@Override
 	public boolean supports(ConversionRequest request) {
-		return request.sourceClass() == int[].class
+
+		return request.sourceClass() == long[].class
 				&& request.destType() == Dimensions.class;
 	}
 }
